This might seem like a small thing, but increases performance quite dramatically in a busy foundry environment.

Now, 'quite dramatically' is still reletive. It will not increase the fps in foundry or make character sheets buttery smooth all of a sudden, but it does greatly reduce UI lag whenever the sounds sidebar is updated, which is at least every time a track is started or completed.

For quite a few people, this will happen directly when combat begins or in other busy situations and in my world with only about 160 sounds distributed among many playlists, just running the `preUpdatePlaylist` moulinette hook takes about 100ms. 100ms itself might not seem bad, but when this is stacked on top of other inefficiencies in other modules, 3 modules each taking about 100ms extra can lead to a sudden, half-second freeze of the ui every now and then, which then is very noticable.

This simple change essentially reduces time this hook runs for me from about 100ms to 6ms.