@SvenWerlen 2024-09-09
That makes sense!
I didn't realize that it would make a difference. I'm sure there are other optimizations like this.
I'll review it within the next few days and merge it.
Thank you for your contribution